+++ CHANGELOG.md
@@ -7,6 +7,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.3] - 2023-02-09
+
+### Fixed
+
+- Mistakes in path prefix extraction when there is only one file in the tooth.
+
+## [0.8.2] - 2023-02-09
+
+### Fixed
+
+- Failing to input anything to post-install and pre-uninstall commands.
+- Wrong installation order of dependencies.
+- Registry not working in `lip show`.
+- Unstable versions can be wrongly installed when no specific version is specified.
+
+## [0.8.1] - 2023-02-07
+
+### Fixed
+
+- Failing to get information from registry with other index than index.json.
+
 ## [0.8.0] - 2023-02-06
 
 ### Added
@@ -131,7 +152,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Basic functions: cache, install, list, show, tooth init, and uninstall.
 
-[unreleased]: https://github.com/LiteLDev/Lip/compare/v0.8.0...HEAD
+[unreleased]: https://github.com/LiteLDev/Lip/compare/v0.8.3...HEAD
+[0.8.3]: https://github.com/LiteLDev/Lip/compare/v0.8.2...v0.8.3
+[0.8.2]: https://github.com/LiteLDev/Lip/compare/v0.8.1...v0.8.2
+[0.8.1]: https://github.com/LiteLDev/Lip/compare/v0.8.0...v0.8.1
 [0.8.0]: https://github.com/LiteLDev/Lip/compare/v0.7.1...v0.8.0
 [0.7.1]: https://github.com/LiteLDev/Lip/compare/v0.7.0...v0.7.1
 [0.7.0]: https://github.com/LiteLDev/Lip/compare/v0.6.0...v0.7.0